+++ js/controllers/editPost.js
@@ -101,12 +101,12 @@ postworld.directive( 'pwEditPost', [ function($scope){
 postworld.controller('editPostController',
 	['$scope', '$rootScope', 'pwPostOptions', 'pwEditPostFilters', '$timeout', '$filter',
 	'embedly', 'pwData', '$log', '$route', '$routeParams', '$location', '$http', '$window',
-	'pwRoleAccess', 'pwQuickEdit', '_', '$sce', 'pwTemplatePartials', 'pwOptionsData', '$pw',
+	'pwRoleAccess', 'pwQuickEdit', '_', '$sce', 'pwTemplatePartials', '$pw',
 	function($scope, $rootScope, $pwPostOptions, $pwEditPostFilters, $timeout, $filter, $embedly,
 		$pwData, $log, $route, $routeParams, $location, $http, $window,
-		$pwRoleAccess, $pwQuickEdit, $_, $sce, $pwTemplatePartials, $pwOptionsData, $pw ) {
+		$pwRoleAccess, $pwQuickEdit, $_, $sce, $pwTemplatePartials, $pw ) {
 
-	$scope['options'] = $pwOptionsData['options'];
+	$scope['options'] = $pw.optionsMeta;
 
 	$log.debug( "$scope.mode @ init", $scope.mode );
 	$log.debug( "$scope.meta @ init", $scope.meta );