+++ templates/admin/select-slider-settings.php
@@ -46,7 +46,8 @@ if( !is_array($show) )
 			));
 		?><hr class="thin"><?php
 	}
-?>
+	?>
+
 <?php
 	///// HEIGHT /////
 	if( in_array( 'height', $show ) ){
@@ -55,15 +56,12 @@ if( !is_array($show) )
 			'ng_model' => $ng_model.'.height',
 			//'methods' => array('window-base','window-percent','pixels'),
 			));
-		?>
-		<hr class="thin"><?php
+		?><hr class="thin"><?php
 	}
 	?>
-	
-<?php
-	///// INTERVAL /////
-	if( in_array( 'interval', $show ) ){
-	?>
+
+<?php if( in_array( 'interval', $show ) ): ?>
+	<!-- INTERVAL -->
 	<span class="icon-md"><i class="pwi-clock"></i></span>
 	<input
 		id="input-interval"
@@ -71,15 +69,14 @@ if( !is_array($show) )
 		size="3"
 		type="number"
 		ng-model="<?php echo $ng_model; ?>.interval">
-		<label for="input-interval"><b><span uib-tooltip="milliseconds">ms</span></b> interval</label>
+		<label for="input-interval">
+			<b><?php _e('milliseconds','postworld') ?></b> <?php _e('interval','postworld') ?>
+		</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// PARALLAX DEPTH /////
-	if( in_array( 'parallax_depth', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'parallax_depth', $show ) ) : ?>
+	<!-- PARALLAX DEPTH -->
 	<span class="icon-md"><i class="pwi-images"></i></span>
 	<input
 		id="parallax-depth"
@@ -87,84 +84,78 @@ if( !is_array($show) )
 		size="3"
 		type="number"
 		ng-model="<?php echo $ng_model; ?>.parallax_depth">
-	<label for="parallax-depth">parallax depth</label>
+	<label for="parallax-depth">
+		<?php _e('parallax depth','postworld') ?>
+	</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// MAXIMUM SLIDES /////
-	if( in_array( 'max_slides', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'max_slides', $show ) ) : ?>
+	<!-- MAXIMUM SLIDES -->
 	<span class="icon-md"><i class="pwi-plus"></i></span>
 	<input type="text"
 		id="input-maxposts"
 		size="3"
 		ng-model="<?php echo $ng_model; ?>.query_vars.max_posts">
-	<label for="input-maxposts">maximum slides</label>
+	<label for="input-maxposts">
+		<?php _e('maximum slides','postworld') ?>
+	</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// HYPERLINK /////
-	if( in_array( 'hyperlink', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'hyperlink', $show ) ) : ?>
+	<!-- HYPERLINK -->
 	<span class="icon-md"><i class="pwi-link"></i></span>
 	<input type="checkbox"
 		id="input-hyperlink"
 		ng-model="<?php echo $ng_model; ?>.hyperlink">
-		<label for="input-hyperlink">Link slides to their respective posts/pages</label>
+		<label for="input-hyperlink">
+			<?php _e('Link slides to their respective posts/pages','postworld') ?>
+		</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// NO PAUSE /////
-	if( in_array( 'no_pause', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'no_pause', $show ) ) : ?>
+	<!-- NO PAUSE -->
 	<span class="icon-md"><i class="pwi-pause"></i></span>
 	<input type="checkbox"
 		id="input-no_pause"
 		ng-model="<?php echo $ng_model; ?>.no_pause">
-		<label for="input-no_pause">Do not pause slider on mouse over</label>
+		<label for="input-no_pause">
+			<?php _e('Do not pause slider on mouse over','postworld') ?>
+		</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// SHOW TITLE /////
-	if( in_array( 'show_title', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'show_title', $show ) ) : ?>
+	<!-- SHOW TITLE -->
 	<span class="icon-md"><i class="pwi-eye"></i></span>
 	<input type="checkbox"
 		id="input-show_title"
 		ng-model="<?php echo $ng_model; ?>.show_title">
-		<label for="input-show_title">Show Title</label>
+		<label for="input-show_title">
+			<?php _e('Show Title','postworld') ?>
+		</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// SHOW EXCERPT /////
-	if( in_array( 'show_excerpt', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'show_excerpt', $show ) ) : ?>
+	<!-- SHOW EXCERPT -->
 	<span class="icon-md"><i class="pwi-eye"></i></span>
 	<input type="checkbox"
 		id="input-show_excerpt"
 		ng-model="<?php echo $ng_model; ?>.show_excerpt">
-		<label for="input-show_excerpt">Show Excerpt</label>
+		<label for="input-show_excerpt">
+			<?php _e('Show Excerpt','postworld') ?>
+		</label>
 	<hr class="thin">
-	<?php
-	}
-?>
-<?php
-	///// TRANSITION /////
-	if( in_array( 'transition', $show ) ){
-	?>
+<?php endif ?>
+
+<?php if( in_array( 'transition', $show ) ) : ?>
+	<!-- TRANSITION -->
 	<label class="inner" for="select-transition">
 		<i class="pwi-wand"></i>
-		Transition
+		<?php _e('Transition','postworld') ?>
 	</label>
 	<select
 		class="labeled"
@@ -173,8 +164,6 @@ if( !is_array($show) )
 		ng-options="option.value as option.name for option in options.transition">
 	</select>
 	<hr class="thin">
-	<?php
-	}
-?>
+<?php endif ?>
 
 </div>